move movie store to its own file
